requirements.txt
streamlit>=1.44.0
openai>=1.68.2
tiktoken>=0.9.0
python-dotenv>=1.0.1
pandas>=2.2.3
plotly>=6.0.1
rank-bm25>=0.2.2
pypdf>=5.4.0
llmlingua>=0.2.2
sentence-transformers>=3.4.1
